+++ management/model_subscription_http_endpoint.go
@@ -20,7 +20,7 @@ var _ MappedNullable = &SubscriptionHttpEndpoint{}
 // SubscriptionHttpEndpoint struct for SubscriptionHttpEndpoint
 type SubscriptionHttpEndpoint struct {
 	// A string that specifies a valid HTTPS URL to which event messages are sent. This is a required property.
-	Url string `json:"url"`
+	Url *string `json:"url,omitempty"`
 	// An object map of strings that specifies the headers applied to the outbound request (for example, `Authorization` `Basic usernamepassword`. The purpose of these headers is for the HTTPS endpoint to authenticate the PingOne service, ensuring that the information from PingOne is from a trusted source.
 	Headers *map[string]string `json:"headers,omitempty"`
 }
@@ -29,9 +29,8 @@ type SubscriptionHttpEndpoint struct {
 // This constructor will assign default values to properties that have it defined,
 // and makes sure properties required by API are set, but the set of arguments
 // will change when the set of required properties is changed
-func NewSubscriptionHttpEndpoint(url string) *SubscriptionHttpEndpoint {
+func NewSubscriptionHttpEndpoint() *SubscriptionHttpEndpoint {
 	this := SubscriptionHttpEndpoint{}
-	this.Url = url
 	return &this
 }
 
@@ -43,28 +42,36 @@ func NewSubscriptionHttpEndpointWithDefaults() *SubscriptionHttpEndpoint {
 	return &this
 }
 
-// GetUrl returns the Url field value
+// GetUrl returns the Url field value if set, zero value otherwise.
 func (o *SubscriptionHttpEndpoint) GetUrl() string {
-	if o == nil {
+	if o == nil || IsNil(o.Url) {
 		var ret string
 		return ret
 	}
-
-	return o.Url
+	return *o.Url
 }
 
-// GetUrlOk returns a tuple with the Url field value
+// GetUrlOk returns a tuple with the Url field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SubscriptionHttpEndpoint) GetUrlOk() (*string, bool) {
-	if o == nil {
+	if o == nil || IsNil(o.Url) {
 		return nil, false
 	}
-	return &o.Url, true
+	return o.Url, true
 }
 
-// SetUrl sets field value
+// HasUrl returns a boolean if a field has been set.
+func (o *SubscriptionHttpEndpoint) HasUrl() bool {
+	if o != nil && !IsNil(o.Url) {
+		return true
+	}
+
+	return false
+}
+
+// SetUrl gets a reference to the given string and assigns it to the Url field.
 func (o *SubscriptionHttpEndpoint) SetUrl(v string) {
-	o.Url = v
+	o.Url = &v
 }
 
 // GetHeaders returns the Headers field value if set, zero value otherwise.
@@ -109,7 +116,9 @@ func (o SubscriptionHttpEndpoint) MarshalJSON() ([]byte, error) {
 
 func (o SubscriptionHttpEndpoint) ToMap() (map[string]interface{}, error) {
 	toSerialize := map[string]interface{}{}
-	toSerialize["url"] = o.Url
+	if !IsNil(o.Url) {
+		toSerialize["url"] = o.Url
+	}
 	if !IsNil(o.Headers) {
 		toSerialize["headers"] = o.Headers
 	}