+++ management/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.68.0 (Unreleased)
+
+* **Enhancement** Added `ConnectionDetails`, `Protocol` to the `Subscription` (webhook) data model.
+* **Enhancement** Added `Tcp` to the `SubscriptionPayloadOptionsPayloadFormatFormat` (webhook) data model.
+* **Enhancement** Made `HttpEndpoint` and `Format` optional in the `Subscription` (webhook) data model.
+
 # v0.67.0 (2026-03-31)
 
 * **Note** Remove redundant data models. [#501](https://github.com/patrickcping/pingone-go-sdk-v2/pull/501)